+++ src/main/java/tech/jhipster/lite/docker/domain/DockerImageName.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.generator.docker.domain;
+package tech.jhipster.lite.docker.domain;
 
 import tech.jhipster.lite.error.domain.Assert;
