+++ src/main/java/tech/jhipster/lite/docker/infrastructure/secondary/FileSystemDockerImagesRepository.java
@@ -1,15 +1,15 @@
-package tech.jhipster.lite.generator.docker.infrastructure.secondary;
+package tech.jhipster.lite.docker.infrastructure.secondary;
 
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.stream.Stream;
 import org.springframework.stereotype.Repository;
 import tech.jhipster.lite.common.domain.ProjectFilesReader;
+import tech.jhipster.lite.docker.domain.DockerImage;
+import tech.jhipster.lite.docker.domain.DockerImageName;
+import tech.jhipster.lite.docker.domain.DockerImages;
+import tech.jhipster.lite.docker.domain.UnknownDockerImageException;
 import tech.jhipster.lite.error.domain.Assert;
-import tech.jhipster.lite.generator.docker.domain.DockerImage;
-import tech.jhipster.lite.generator.docker.domain.DockerImageName;
-import tech.jhipster.lite.generator.docker.domain.DockerImages;
-import tech.jhipster.lite.generator.docker.domain.UnknownDockerImageException;
 
 @Repository
 class FileSystemDockerImagesRepository implements DockerImages {
