+++ src/main/java/tech/jhipster/lite/generator/server/springboot/database/mariadb/application/MariaDBApplicationService.java
@@ -1,7 +1,7 @@
 package tech.jhipster.lite.generator.server.springboot.database.mariadb.application;
 
 import org.springframework.stereotype.Service;
-import tech.jhipster.lite.generator.docker.domain.DockerImages;
+import tech.jhipster.lite.docker.domain.DockerImages;
 import tech.jhipster.lite.generator.server.springboot.database.mariadb.domain.MariaDBModuleFactory;
 import tech.jhipster.lite.module.domain.JHipsterModule;
 import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
