+++ src/main/java/tech/jhipster/lite/generator/server/springboot/database/mongodb/domain/MongodbDomainService.java
@@ -7,9 +7,9 @@ import static tech.jhipster.lite.generator.server.springboot.database.mongodb.do
 
 import java.util.Map;
 import java.util.TreeMap;
+import tech.jhipster.lite.docker.domain.DockerImages;
 import tech.jhipster.lite.generator.buildtool.generic.domain.BuildToolService;
 import tech.jhipster.lite.generator.buildtool.generic.domain.Dependency;
-import tech.jhipster.lite.generator.docker.domain.DockerImages;
 import tech.jhipster.lite.generator.project.domain.DatabaseType;
 import tech.jhipster.lite.generator.project.domain.Project;
 import tech.jhipster.lite.generator.project.domain.ProjectFile;
