+++ src/main/java/tech/jhipster/lite/generator/server/springboot/database/mysql/application/MySQLApplicationService.java
@@ -1,7 +1,7 @@
 package tech.jhipster.lite.generator.server.springboot.database.mysql.application;
 
 import org.springframework.stereotype.Service;
-import tech.jhipster.lite.generator.docker.domain.DockerImages;
+import tech.jhipster.lite.docker.domain.DockerImages;
 import tech.jhipster.lite.generator.server.springboot.database.mysql.domain.MySQLModuleFactory;
 import tech.jhipster.lite.module.domain.JHipsterModule;
 import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
