+++ src/main/java/tech/jhipster/lite/generator/server/springboot/database/postgresql/application/PostgresqlApplicationService.java
@@ -1,7 +1,7 @@
 package tech.jhipster.lite.generator.server.springboot.database.postgresql.application;
 
 import org.springframework.stereotype.Service;
-import tech.jhipster.lite.generator.docker.domain.DockerImages;
+import tech.jhipster.lite.docker.domain.DockerImages;
 import tech.jhipster.lite.generator.server.springboot.database.postgresql.domain.PostgresqlModuleFactory;
 import tech.jhipster.lite.module.domain.JHipsterModule;
 import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
