+++ src/main/java/tech/jhipster/lite/generator/server/springboot/database/sqlcommon/domain/SQLCommonModuleBuilder.java
@@ -4,8 +4,8 @@ import static tech.jhipster.lite.common.domain.WordUtils.*;
 import static tech.jhipster.lite.generator.server.springboot.core.domain.SpringBoot.*;
 import static tech.jhipster.lite.module.domain.JHipsterModule.*;
 
+import tech.jhipster.lite.docker.domain.DockerImage;
 import tech.jhipster.lite.error.domain.Assert;
-import tech.jhipster.lite.generator.docker.domain.DockerImage;
 import tech.jhipster.lite.generator.project.domain.DatabaseType;
 import tech.jhipster.lite.generator.server.springboot.common.domain.Level;
 import tech.jhipster.lite.module.domain.DocumentationTitle;
