+++ src/main/java/tech/jhipster/lite/generator/server/springboot/mvc/security/oauth2/application/OAuth2SecurityApplicationService.java
@@ -1,7 +1,7 @@
 package tech.jhipster.lite.generator.server.springboot.mvc.security.oauth2.application;
 
 import org.springframework.stereotype.Service;
-import tech.jhipster.lite.generator.docker.domain.DockerImages;
+import tech.jhipster.lite.docker.domain.DockerImages;
 import tech.jhipster.lite.generator.server.springboot.mvc.security.oauth2.domain.OAuth2AccountModuleFactory;
 import tech.jhipster.lite.generator.server.springboot.mvc.security.oauth2.domain.OAuth2ModuleFactory;
 import tech.jhipster.lite.module.domain.JHipsterModule;
