+++ src/main/java/tech/jhipster/lite/generator/server/springboot/springcloud/common/domain/SpringCloudCommonDomainService.java
@@ -9,9 +9,9 @@ import java.io.IOException;
 import java.util.List;
 import tech.jhipster.lite.common.domain.Base64Utils;
 import tech.jhipster.lite.common.domain.FileUtils;
+import tech.jhipster.lite.docker.domain.DockerImages;
 import tech.jhipster.lite.error.domain.GeneratorException;
 import tech.jhipster.lite.generator.buildtool.generic.domain.BuildToolService;
-import tech.jhipster.lite.generator.docker.domain.DockerImages;
 import tech.jhipster.lite.generator.project.domain.Project;
 import tech.jhipster.lite.generator.project.domain.ProjectFile;
 import tech.jhipster.lite.generator.project.domain.ProjectRepository;
