+++ src/main/java/tech/jhipster/lite/generator/server/springboot/springcloud/common/infrastructure/SpringCloudCommonBeanConfiguration.java
@@ -2,8 +2,8 @@ package tech.jhipster.lite.generator.server.springboot.springcloud.common.infras
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import tech.jhipster.lite.docker.domain.DockerImages;
 import tech.jhipster.lite.generator.buildtool.generic.domain.BuildToolService;
-import tech.jhipster.lite.generator.docker.domain.DockerImages;
 import tech.jhipster.lite.generator.project.domain.ProjectRepository;
 import tech.jhipster.lite.generator.server.springboot.springcloud.common.domain.SpringCloudCommonDomainService;
 import tech.jhipster.lite.generator.server.springboot.springcloud.common.domain.SpringCloudCommonService;
